services/agent-service/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
langchain-google-genai
google-generativeai
pydantic
pytest
pytest-asyncio
httpx
pytest-mock
requests-mock
